fix: handle sort order conversion
